Creating Oracle related project stubs
The following projects have been created:
Ora.Base - base wrapper for connecting to database, enumerating data providers and for other similar stuff.
Ora.Cache - this is the target project of data collection that has been changed in database;
Ora.Notify - this is the main tool for getting notifications from database on events of data change.